requirements.txt
PyQt6==6.5.0
Pillow==9.4.0
toml==0.10.2